Rename base model to pytorch and simplify methods
