+++ nilm_thresholding/model/gru.py
@@ -2,7 +2,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 import torch.optim as optim
 
-from nilm_thresholding.model.base import TorchModel
+from nilm_thresholding.model.pytorch import TorchModel
 
 
 class _Dense(nn.Module):
